+++ crates/aptos/CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to the Aptos CLI will be captured in this file. This project
 ## [4.5.0] - 2024/11/15
 - Determine network from URL to make explorer links better for legacy users
 - Add support for AIP-80 compliant strings when importing using the CLI arguments or manual input.
+- Add option `--print-compiler-version-only` to `aptos move decompile` and `aptos move disassemble` to print out the version of compiler that generated the target bytecode.
 
 ## [4.4.0] - 2024/11/06
 - Fix typos in `aptos move compile` help text.